added support to jakartaee in jsf plugin
