+++ apm-agent-plugins/apm-jsf-plugin/pom.xml
@@ -22,5 +22,11 @@
             <version>7.0</version>
             <scope>provided</scope>
         </dependency>
+        <dependency>
+            <groupId>jakarta.platform</groupId>
+            <artifactId>jakarta.jakartaee-api</artifactId>
+            <version>9.0.0</version>
+            <scope>provided</scope>
+        </dependency>
     </dependencies>
 </project>
