+++ apm-agent-plugins/apm-jsf-plugin/src/main/resources/META-INF/services/co.elastic.apm.agent.sdk.ElasticApmInstrumentation
@@ -1,2 +1,2 @@
-co.elastic.apm.agent.jsf.JsfLifecycleInstrumentation$JsfLifecycleExecuteInstrumentation
-co.elastic.apm.agent.jsf.JsfLifecycleInstrumentation$JsfLifecycleRenderInstrumentation
+co.elastic.apm.agent.jsf.JsfLifecycleExecuteInstrumentation
+co.elastic.apm.agent.jsf.JsfLifecycleRenderInstrumentation
